% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eForest.R
\name{eForest}
\alias{eForest}
\title{Autoencoder forest}
\usage{
eForest(rf, x, parallel = TRUE)
}
\arguments{
\item{rf}{Random forest model object of class \code{ranger}.}

\item{x}{Input data. May be the training data used for \code{rf}, or else
an independent test set.}

\item{parallel}{Compute in parallel? Must register backend beforehand.}
}
\value{
An encoded dataset of maximum compatible rules for each sample.
}
\description{
Implements an autoencoder forest, a function that maps inputs to a
hyperrectangle defined by the intersection of leaves in a random forest.
}
\details{
This function implements the autoencoder forest algorithm proposed by Feng &
Zhou (2017).
}
\references{
Feng, J. and Zhou, Z. (2017). AutoEncoder by Forest. \emph{arXiv} preprint,
1709.09018.
}
