% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eForestSynth.R
\name{eForestSynth}
\alias{eForestSynth}
\title{Synthesize data from autoencoder forest}
\usage{
eForestSynth(rf, x, z, n_synth)
}
\arguments{
\item{rf}{Random forest model object of class \code{ranger}.}

\item{x}{Input data. May be the training data used for \code{rf}, or else
an independent test set.}

\item{z}{Autoencoded data as computed by \code{eForest}.}

\item{nSynth}{Number of synthetic samples to draw.}
}
\value{
A dataset of synthetic samples.
}
\description{
Takes uniform draws from hyperrectangles computed by an autoencoder forest.
}
\details{
This function creates synthetic samples from an autoencoder forest.
}
\references{
Feng, J. and Zhou, Z. (2017). AutoEncoder by Forest. \emph{arXiv} preprint,
1709.09018.
}
